server/requirements.txt
# Flask and Extensions
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.1.1

# Authentication and Environment
python-dotenv==1.0.0
Werkzeug==2.3.7

# HTTP Requests
requests==2.31.0

# Database and Data Processing
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Data Science and Recommendations
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# Utilities
python-dateutil==2.8.2
click==8.1.7
